change float into int

+++ main.py
@@ -1,3 +1,3 @@
 temperature = input("Enter temperature in celsius: ")
-temperature = int(temperature)
+temperature = float(temperature)
 print(f"{temperature}° in Celsius is equivalent to {temperature *  9/5 + 32}° Fahrenheit. ")
